Separate slugify_in_context into util.py
