+++ substanced/folder/views.py
@@ -6,7 +6,6 @@ import itertools
 import operator
 import re
 import os
-import unidecode
 import colander
 import venusian
 
@@ -36,27 +35,12 @@ from ..sdi import (
     sdi_mgmt_views,
     RIGHT,
     )
+from .util import slugify_in_context
 from ..util import _
 
 from . import FolderKeyError
 
 _marker = object()
-
-re_word = re.compile(r'\W+')
-
-def slugify_in_context(context, name, remove_extension=True):
-    if remove_extension:
-        name = os.path.splitext(name)[0]
-
-    slug = unidecode.unidecode(name).lower()
-    orig = slug = re_word.sub('-', slug)
-    i = 1
-    while True:
-        if slug not in context:
-            break
-        slug = '%s-%i' % (orig, i)
-        i += 1
-    return slug
 
 
 class folder_contents_views(object):
